paper-58-MindGlitch/requirements.txt
# Core ML libraries (PyTorch with CUDA - install separately)
transformers
omegaconf
huggingface_hub
safetensors
datasets
scikit-learn

# Computer Vision & Image Processing
opencv-python
pillow
matplotlib
scipy
numpy

# Diffusion Models & Inpainting
diffusers
accelerate

# Feature Extraction & Correspondence
einops
jaxtyping

# Metrics & Evaluation
lpips
openai-clip

# Segmentation Models
segment-anything

# Configuration & Utilities
hydra-core
python-dotenv
tqdm
wandb

# Detectron2 (for ResNet backbone)
fvcore

# Additional utilities
GitPython
ipykernel